Moved queue class
