+++ core/lib/index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="./animations.css" />
     <script type="module" src="/core/pb/dist/pocketbase.umd.js"></script>
     <script type="module" src="./main.js"></script>
+    <script type="module" src="./queue.js"></script>
     <script type="module" src="./notifs.ts"></script>
     <title>Operation Grace</title>
   </head>
